requirements.txt
fastapi
uvicorn
python-dotenv
pinecone
langchain
langchain-core
langchain-community
langchain-openai
langchain-pinecone
tiktoken
pandas
sentence-transformers